refactor(arch/*/kernel): factor out duplicated code
