+++ src/arch/riscv64/kernel/core_local.rs
@@ -69,21 +69,6 @@ impl CoreLocal {
 	}
 }
 
-#[inline]
-pub fn core_id() -> CoreId {
-	CoreLocal::get().core_id
-}
-
-#[inline]
-pub fn core_scheduler() -> &'static mut PerCoreScheduler {
-	unsafe { CoreLocal::get().scheduler.get().as_mut().unwrap() }
-}
-
-#[inline]
-pub fn set_core_scheduler(scheduler: *mut PerCoreScheduler) {
-	CoreLocal::get().scheduler.set(scheduler);
-}
-
-pub(crate) fn ex() -> &'static StaticExecutor<RawSpinMutex, RawRwSpinLock> {
-	&CoreLocal::get().ex
-}
+#[path = "../../../kernel/core_local.rs"]
+mod core_local_common;
+pub use core_local_common::*;
