+++ src/kernel/README.md
@@ -1,0 +1,3 @@
+# kernel
+
+This directory is used to collect architecture-unspecific code used in `src/arch/*/kernel`.
